+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added command to get information about Network Interfaces
+
+### Updated
+
+- Updated Zen command to show a fidglet of cli name
+
 ## [0.0.1] - 2021-06-10
 
 ### Added